Force of infection equations

FoI combines parameters related to HIV transmission via needle sharing, heterosexual sex and homosexual sex

FoI(
  y,
  no,
  uoC,
  ns,
  usC,
  eO,
  eS,
  sigmaFM,
  sigmaMF,
  sigmaM,
  tau,
  eff.prep,
  d,
  s,
  eff.oat,
  cov.ssp,
  eff.ssp = eff.ssp,
  s.multi,
  bal = F
)

Arguments

y
number of individuals in each compartment (n.gp*19 matrix)

no
number of opposite sexual partners for each compartment (n.gp*19 matrix)

uoC
condom use adjustment term for opposite sex (n.gp*19 matrix)

ns
number of same sexual partners for each compartment (18*19 matrix)

usC
condom use adjustment term for same sex (18*19 matrix)

eO
assortative coefficient for heterosexual mixing between ethnic groups (vector of 3 for white, black, hispanic)

eS
assortative coefficient for homosexual mixing between ethnic groups (vector of 3 for white, black, hispanic)

sigmaFM
probability of transmission by F->M for 16 HIV+ states

sigmaMF
probability of transmission by M->F for 16 HIV+ states

tau
probability of transmission by injection for 16 HIV+ states

eff.prep
% reduction in risk of infection for PrEP

d
number of injection

s
proportion of injections that are shared (9 groups)

eff.oat
% reduction in # of shared injections reduced due to OAT

cov.ssp
coverage of syringe services program

simgaM
probability of transmission by same sex for 16 HIV+ states

outputs
matrix of dim c(n.gp,2) with row of each group. second column for PrEP

ind
index for groups of people to be mixed

ind.low
index for groups of people at low risk of HIV transmission

ass.e
Assortativity coefficient

n
Number of contacts

uC
condom use matrix

indM
indicator of male partner

indF
indicator of female partner

Value

Sufficient contact rate (sum of transmission via needle sharing, opposite and sex same contacts)

Complete mixing matrix