Specify ODE equations for PWID off OAT

ode.list.offOAT Calculate the change of population in cells of PWID population but off OAT by the ODE model

ode.list.offOAT(
  x,
  x.oat,
  beta_o,
  beta_s,
  gamma,
  rho,
  ws,
  wp,
  mu_mat,
  mo,
  eta,
  psi,
  psi.p,
  theta.ai,
  theta.ad,
  phi,
  v2,
  v3,
  alpha,
  alpha.re,
  theta.1,
  theta.2,
  theta.t12,
  theta.t13,
  theta.t21,
  theta.t23,
  theta.t31,
  theta.t32,
  theta.t1O,
  theta.t2O,
  theta.t3O,
  oat.e,
  oat.q,
  rho.m
)

Arguments

x
a vector containing the number of individuals off OAT x = c(S1,S2,Sp,Ia,I1,I2,I3,Iap,Ip,Da,D1,D2,D3,T1,T2,T3,O1,O2,O3)

x.oat
a vector containing the number of individuals on OAT

rho
the entry rate (assume the same for all groups)

mu_mat
maturation rates (age>65)

mo
mortality rates for 19 states

eta
PrEP entry rate

psi
screening rates for susceptible or infected

psi.p
screening rates for Iap (acute HIV on PrEP)

theta.ai
transition rate from acute states (Ia) to chronic state (CD4>=500; I1)

theta.ad
transition rate from acute states (Da) to chronic state (CD4>=500; D1)

phi
% infected of people receiving ART once diagnosed

v2, v3
symptom-based case finding rate for I2 and I3, respectively

alpha
ART initiation rate for D1,D2, and D3

alpha.re
ART re-initiation rate for o1,O2, and O3

theta.1
HIV disease progression rate for those not on ART, from I1/D1 to I2/D2

theta.2
HIV disease progression rate for those not on ART, from I2/D2 to I3/D3

oat.e
OAT entry rate

oat.q
OAT dropout rate

lambda
the sufficient contact rate

lambda.p
the sufficient contact rate for those on PrEP.

1/ws
average duration uninfected individuals in compartment S remain identified after screening

1/wp
average duration uninfected individuals in compartment Sp remain on PrEP

psi.pr
entry rate for PrEP

theta.t
transition probabilities for those on ART

theta.o
ART dropout probability from states T1,T2,T3

Value

HIV diagnoses, incidence and deaths for PWID off OAT